Rename ExecutionEngineService => ExecutionEngineChannelImpl. Add ExecutionEngineService and incorporate events listening logic there
