+++ services/powchain/build.gradle
@@ -8,6 +8,7 @@ dependencies {
   implementation project(':infrastructure:logging')
   implementation project(':infrastructure:version')
   implementation project(':pow')
+  implementation project(':storage')
   implementation project(':storage:api')
   implementation project(':services:serviceutils')
   implementation project(':util')
